+++ tests/functional/eyes_selenium/selenium/test_specific_cases.py
@@ -239,3 +239,19 @@ def test_feature_target_window_captures_selected_frame(eyes, driver):
 
     eyes.check("step name", Target.window())
     eyes.close()
+
+
+def test_capture_element_on_pre_scrolled_down_page(eyes, driver):
+    driver.get(
+        "http://applitools.github.io/demo/TestPages/FramesTestPage/longframe.html"
+    )
+    eyes.open(
+        driver=driver,
+        app_name="Applitools Eyes SDK",
+        test_name="Test capture element on pre scrolled down page",
+        viewport_size={"width": 800, "height": 600},
+    )
+    driver.execute_script("window.scrollTo(0, 300)")
+    eyes.check("Row 10", Target.region("body > table > tr:nth-child(10)"))
+    eyes.check("Row 20", Target.region("body > table > tr:nth-child(20)"))
+    eyes.close()